+++ src/main/java/net/zuiron/photosynthesis/recipe/CropSticksRecipe.java
@@ -38,7 +38,8 @@ public class CropSticksRecipe implements Recipe<SimpleInventory> {
 
         return recipeItems.get(0).test(inventory.getStack(0))
                 && recipeItems.get(1).test(inventory.getStack(1))
-                && recipeItems.get(2).test(inventory.getStack(2));
+                && recipeItems.get(2).test(inventory.getStack(2))
+                && recipeItems.get(3).test(inventory.getStack(3));
     }
 
     @Override